+++ include/substances/Argon.hpp
@@ -9,7 +9,7 @@
 
 #define SHARPNESS_COEFFICIENT_AR 100.0
 
-Enthalpy<double> internal_energy_gaseous_Ar(
+Enthalpy<long double> internal_energy_gaseous_Ar(
     20.78600,
     2.825911E-7,
    -1.464191E-7,
@@ -18,9 +18,9 @@ Enthalpy<double> internal_energy_gaseous_Ar(
    -6.197350
 );
 
-ThermalConductivityQuadraticPolynomial<double> thermal_conductivity_gaseous_Ar(1.49E-3, 5.98E-5, -1.92E-8);
+ThermalConductivityQuadraticPolynomial<long double> thermal_conductivity_gaseous_Ar(1.49E-3, 5.98E-5, -1.92E-8);
 
-Phase<double> gaseous_Ar(
+Phase<long double> gaseous_Ar(
     1.6025,
     internal_energy_gaseous_Ar,
     thermal_conductivity_gaseous_Ar,
@@ -29,8 +29,8 @@ Phase<double> gaseous_Ar(
     SHARPNESS_COEFFICIENT_AR
 );
 
-Phase<double> phases_Ar[] = {gaseous_Ar};
+Phase<long double> phases_Ar[] = {gaseous_Ar};
 
-Substance<double> Argon(1, phases_Ar, 39.948E-3, 5.0/3.0);
+Substance<long double> Argon(1, phases_Ar, 39.948E-3, 5.0/3.0);
 
 #endif
